nothing found

+++ app/models/user.py
@@ -1,4 +1,6 @@
 from app import db
+# from sqlalchemy import Numeric
+# from decimal import Decimal
 
 class User(db.Model):
     user_id = db.Column(db.Integer, primary_key=True, autoincrement = True)
@@ -6,7 +8,8 @@ class User(db.Model):
     name = db.Column(db.String, nullable=False)
     email = db.Column(db.String, nullable=False)
     zip_code = db.Column(db.Integer, nullable=False)
-    tennis_level = db.Column(db.Float)
+    tennis_level = db.Column(db.Integer)
+    # tennis_level = db.Column(Numeric(precision=10, scale=2,decimal_return_scale=None))
     preferences = db.Column(db.String(300))
 
     def to_dict(self):
